pull updates into forked main

+++ controllers/api/item-routes.js
@@ -1,30 +1,39 @@
-const router = require('express').Router();
-const {List, Item} = require('../../models');
+const router = require("express").Router();
+const { List, Item } = require("../../models");
 
 // api/items endpoint
 
-router.get('/', async (req, res) => {
-    try {
-        const itemData = await Item.findAll({
-            include: [{model: List}]
-        });
-        res.status(200).json(itemData);
-    } catch (err) {
-        res.status(500).json(err);
-    }
-});
+// get on all items disabled. Enable only for testing.
+/*
+  router.get('/', async (req, res) => {
+      try {
+          const itemData = await Item.findAll({
+              include: [{model: List}]
+          });
+          res.status(200).json(itemData);
+      } catch (err) {
+          res.status(500).json(err);
+      }
+  });
+*/
 
-router.get('/:id', async (req,res) => {
-    try {
-        const itemData = await Item.findByPk(req.params.id, {
-            include: [{model: List}]
-        });
+router.get("/:id", async (req, res) => {
+  try {
+    const itemData = await Item.findByPk(req.params.id, {
+      include: [{ model: List }],
+    });
 
         if(!itemData) {
             res.status(404).json({message: 'No item with that ID.'});
             return;
         }
-        res.status(200).json(itemData);
+        const serializedItem = itemData.get({plain: true});
+        
+        if(serializedItem.list.user_id !== req.session.userID) {
+          res.status(401).json({message: "This item is not part of one of your lists. Please log in as the correct user."});
+          return;
+        }
+        res.status(200).json(serializedItem);
     } catch (err) {
         res.status(500).json(err);
     }
@@ -32,6 +41,11 @@ router.get('/:id', async (req,res) => {
 
 router.post('/', async (req, res) => {
     try {
+      const itemsList = await List.findByPk(req.body.list_id);
+      if(itemsList.user_id !== req.session.userID) {
+        res.status(401).json({message: "You don't own the list you're trying to add to. Please log in to the correct user."});
+        return;
+      }
       const itemData = await Item.create(req.body);
       res.status(200).json(itemData);
     } catch (err) {
@@ -42,17 +56,50 @@ router.post('/', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
     try {
-      const itemData = await Item.destroy({
-        where: {id: req.params.id}
+      const itemData = await Item.findByPk(req.params.id, {
+        include: [{model: List}]
       });
       if(!itemData) {
-        res.status(404).json({message: 'no tag with that id.'});
+        res.status(404).json({message: 'no item with that id.'});
         return;
       }
-      res.status(200).json(itemData);
+      
+      const serializedItem = itemData.get({plain: true});
+      if(serializedItem.list.user_id !== req.session.userID) {
+        res.status(401).json({message: "This item is not part of one of your lists. Please log in as the correct user."});
+        return;
+      }
+      const destroyItem = await Item.destroy({
+        where: {id: req.params.id}
+      });
+      res.status(200).json(destroyItem);
     } catch (err) {
       res.status(500).json(err);
     }
-  });
+});
+
+// Update items for claim function
+router.put("/:id", async (req, res) => {
+  try {
+    // Check if user claiming or unclaiming the item is the logged in user
+    if (req.body.user_id != req.session.userID && req.body.user_id !== null) {
+      req.status(401).json({
+        message: "You are not the user claiming this item. Please log in as the correct user.",
+      });
+      return;
+    }
+    
+    // Note: add verification to the model
+    if (req.body.status) {
+      const itemData = await Item.update(
+        { status: req.body.status, claimed_user: req.body.user_id },
+        { where: { id: req.params.id } }
+      );
+      res.status(200).json(itemData);
+    }
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
 
 module.exports = router;
